rename package, no gap if available flag passed
