+++ cli/internal/pkg/middleware/scdr/response.go
@@ -5,6 +5,7 @@ import(
   "strconv"
   "time"
   "github.com/cedardevs/onestop/cli/internal/pkg/flags"
+  "fmt"
 )
 
 func MarshalScdrResponse(params *viper.Viper, data interface{}) interface{} {
@@ -21,9 +22,11 @@ func transformResponse(params *viper.Viper, responseMap map[string]interface{}) 
 		isSummary := params.GetString(flags.AvailableFlag)
 		gapInterval := params.GetString(flags.GapFlag)
     typeArg := params.GetString(flags.TypeFlag)
+    fmt.Println(typeArg)
+    fmt.Println(isSummary)
 
     //gap is ignored if no type is passed or if --available is passed
-		if len(gapInterval) > 0 && len(typeArg) > 0 && len(isSummary) == 0 {
+		if len(gapInterval) > 0 && len(typeArg) > 0 && isSummary ==  "false"{
 			scdrOuput = FindGaps(gapInterval, items)
 		} else if isSummary == "true" {
 			count := getCount(responseMap)
